# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 1.0.0
  description: |
    ClinicFlow REST contract. Sessions resolve a database role. The HttpOnly, Secure,
    SameSite=Lax cookie contains an opaque native token by default for continuity,
    or an app-issued JWT when AUTH_SESSION_MODE=jwt is explicitly configured.
    Both modes are database-revocable with a fixed 12-hour lifetime, with no refresh
    endpoint. Changing modes invalidates old cookies. JWT_SIGNING_KEY must be a
    dedicated high-entropy secret of at least 32 UTF-8 bytes, configured before
    enabling JWT mode. Session issuance may return 503 AUTH_SESSION_UNCONFIGURED.
    Newly created passwords require at least 8 characters, ASCII letters and numbers,
    and at most 1024 UTF-8 bytes (enforced by the server).
    Super admins access all; clinic admins and receptionists are assignment-scoped;
    doctors access their own resources and relevant patients; patients access self only.
    All filters narrow authorized scope, never expand it. DELETE deactivates records
    with history. Lists contain real persisted data only. Queue clients poll every 30 seconds.
    Invalid input returns 400, unauthenticated 401, forbidden 403, missing 404,
    state/capacity conflicts 409, rate limiting 429, unavailable integrations 503.
servers:
  - url: /api
security:
  - NativeSession: []
paths:
...
  /clinic-registration:
    post:
      operationId: registerClinic
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/ClinicRegistrationInput" } } } }
      responses:
        "201": { description: Atomic clinic registration, content: { application/json: { schema: { $ref: "#/components/schemas/ClinicSettingsResult" } } } }
  /clinics/{id}/settings:
    parameters:
      - { $ref: "#/components/parameters/Id" }
    get:
      operationId: getClinicSettings
      responses:
        "200": { description: Clinic settings, content: { application/json: { schema: { $ref: "#/components/schemas/ClinicSettingsResult" } } } }
    patch:
      operationId: updateClinicSettings
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/ClinicSettingsInput" } } } }
      responses:
        "200": { description: Saved atomically, content: { application/json: { schema: { $ref: "#/components/schemas/ClinicSettingsResult" } } } }
  /clinics/{id}/settings/preview:
    parameters:
      - { $ref: "#/components/parameters/Id" }
    post:
      operationId: previewClinicSettings
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/ClinicSettingsInput" } } } }
      responses:
        "200": { description: Advisory linked schedule impact, content: { application/json: { schema: { $ref: "#/components/schemas/ClinicSettingsPreview" } } } }
  /me/doctor-profile:
    post:
      operationId: attachOwnDoctorProfile
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/OwnDoctorProfileInput" } } } }
      responses:
        "200": { description: Own clinical capability, content: { application/json: { schema: { $ref: "#/components/schemas/Doctor" } } } }
  /public/slug-availability:
    get:
      operationId: checkSlugAvailability
      security: []
      parameters:
        - { name: slug, in: query, required: true, schema: { type: string, minLength: 3, maxLength: 63 } }
        - { name: clinicId, in: query, schema: { type: string } }
      responses:
        "200": { description: Advisory availability, content: { application/json: { schema: { $ref: "#/components/schemas/SlugAvailability" } } } }
...
components:
  securitySchemes:
    NativeSession:
      type: apiKey
      in: cookie
      name: digiq_session
      description: Opaque native token (default) or explicitly enabled app-issued JWT. Both are database-revocable, fixed 12-hour HttpOnly/Secure/SameSite=Lax cookies. No refresh endpoint.
...
  parameters:
...
    Id: { name: id, in: path, required: true, schema: { type: string } }
...
  schemas:
...
    SlugAvailability:
      type: object
      required: [available, slug]
      properties:
        available: { type: boolean }
        slug: { type: string }
    OpeningHour:
      type: object
      required: [dayOfWeek, startTime, endTime]
      properties:
        dayOfWeek: { type: integer, minimum: 0, maximum: 6 }
        startTime: { type: string }
        endTime: { type: string }
    ClinicPolicy:
      type: object
      properties:
        bookingHorizonDays: { type: integer, minimum: 1, maximum: 365 }
        cancellationCutoffMinutes: { type: integer, minimum: 0, maximum: 10080 }
    ClinicDetails:
      type: object
      properties:
        dateFormat: { $ref: "#/components/schemas/ClinicDateFormat" }
        timeFormat: { $ref: "#/components/schemas/ClinicTimeFormat" }
        name: { type: string, minLength: 1 }
        address: { type: string }
        email: { type: [string, "null"], format: email }
        phone: { type: [string, "null"] }
        slug: { type: string, minLength: 3, maxLength: 63 }
        categoryId: { type: [string, "null"] }
        specialityIds: { type: array, items: { type: string } }
        referralCode: { type: [string, "null"], maxLength: 100 }
    ClinicBranchSetup:
      type: object
      required: [name, address]
      properties:
        id: { type: string }
        name: { type: string, minLength: 1 }
        address: { type: string }
        city: { type: string }
        timezone: { type: string }
        email: { type: [string, "null"], format: email }
        phone: { type: [string, "null"] }
        inheritEmail: { type: boolean }
        inheritPhone: { type: boolean }
        slug: { type: string, minLength: 3, maxLength: 63 }
        openingHours: { type: [array, "null"], maxItems: 28, description: Null or absent preserves legacy unrestricted hours. An explicit empty array closes all days., items: { $ref: "#/components/schemas/OpeningHour" } }
        linkedSchedule: { $ref: "#/components/schemas/LinkedSchedule" }
    OwnerSchedule:
      type: object
      required: [maxTokens, consultationMinutes, tokenPrefix, queueMode]
      properties:
        maxTokens: { type: integer, minimum: 1, maximum: 1000 }
        consultationMinutes: { type: integer, minimum: 1, maximum: 240 }
        tokenPrefix: { type: string, pattern: "^[A-Za-z0-9]{1,8}$" }
        queueMode: { type: string, enum: [mixed, appointmentsOnly, walkInsOnly] }
    LinkedSchedule:
      type: object
      required: [enabled]
      properties:
        enabled: { type: boolean }
        doctorId: { type: string }
        maxTokens: { type: integer, minimum: 1, maximum: 1000 }
        consultationMinutes: { type: integer, minimum: 1, maximum: 240 }
        tokenPrefix: { type: string, pattern: "^[A-Za-z0-9]{1,8}$" }
        queueMode: { type: string, enum: [mixed, appointmentsOnly, walkInsOnly] }
    ClinicSettingsPreview:
      type: object
      required: [allowed, impacts, conflicts]
      properties:
        allowed: { type: boolean }
        conflicts: { type: array, items: { type: string } }
        impacts:
          type: array
          items:
            type: object
            required: [branchId, create, update, retire, unlink]
            properties:
              branchId: { type: string }
              create: { type: integer }
              update: { type: integer }
              retire: { type: integer }
              unlink: { type: boolean }
    ClinicSettingsInput:
      type: object
      properties:
        clinic: { $ref: "#/components/schemas/ClinicDetails" }
        branches: { type: array, maxItems: 30, items: { $ref: "#/components/schemas/ClinicBranchSetup" } }
        policies: { $ref: "#/components/schemas/ClinicPolicy" }
    ClinicSettingsResult:
      type: object
      required: [clinic, branches, policies]
      properties:
        clinic: { $ref: "#/components/schemas/Clinic" }
        branches: { type: array, items: { $ref: "#/components/schemas/Branch" } }
        policies: { $ref: "#/components/schemas/ClinicPolicy" }
        doctorId: { type: [string, "null"] }
    OwnDoctorProfileInput:
      type: object
      required: [branchIds]
      properties:
        branchIds: { type: array, maxItems: 30, items: { type: string } }
        specializationId: { type: string }
        qualificationIds: { type: array, items: { type: string } }
        about: { type: string }
    ClinicRegistrationInput:
      type: object
      required: [fullName, password, clinic, branches]
      properties:
        fullName: { type: string, minLength: 1, maxLength: 150 }
        mobile: { type: string }
        password: { type: string, minLength: 1, maxLength: 1024, writeOnly: true }
        clinic: { $ref: "#/components/schemas/ClinicDetails" }
        branches: { type: array, minItems: 1, maxItems: 30, items: { $ref: "#/components/schemas/ClinicBranchSetup" } }
        policies: { $ref: "#/components/schemas/ClinicPolicy" }
        ownDoctor: { type: boolean, default: false }
        ownerSchedule: { $ref: "#/components/schemas/OwnerSchedule" }
        specializationId: { type: string }
        qualificationIds: { type: array, items: { type: string } }
...
    RecordStatus: { type: string, enum: [active, inactive] }
...
    ClinicDateFormat:
      type: string
      enum: ["DD MMM YYYY", "DD/MM/YYYY", "MM/DD/YYYY", "YYYY-MM-DD"]
      description: Parent Clinic Group display preference; locations inherit it. Default DD MMM YYYY.
    ClinicTimeFormat:
      type: string
      enum: ["12h", "24h"]
      description: Parent Clinic Group display preference; locations inherit it. Default 12h.
    ClinicInput:
      type: object
      required: [name, address]
      properties:
        dateFormat: { $ref: "#/components/schemas/ClinicDateFormat" }
        timeFormat: { $ref: "#/components/schemas/ClinicTimeFormat" }
        slug: { type: string, minLength: 3, maxLength: 63 }
        specialityIds: { type: array, items: { type: string } }
        referralCode: { type: [string, "null"], maxLength: 100 }
        adminId: { type: string, description: Super admins may select the owning clinic administrator; other roles are assigned by the server. }
        name: { type: string, minLength: 1 }
        address: { type: string }
        country: { type: string }
        state: { type: string }
        city: { type: string }
        area: { type: string }
        pincode: { type: string }
        phone: { type: [string, "null"] }
        email: { type: [string, "null"], format: email }
        description: { type: string }
        clinicTypeId: { type: string }
        categoryId: { type: [string, "null"] }
        status: { $ref: "#/components/schemas/RecordStatus" }
    Clinic:
      allOf:
        - { $ref: "#/components/schemas/ClinicInput" }
        - type: object
          required: [id, code, adminId, adminName, status, createdAt]
          properties:
            id: { type: string }
            code: { type: string }
            adminName: { type: string }
            createdAt: { type: [string, "null"], format: date-time }
            doctorCount: { type: integer }
            branchCount: { type: integer }
    BranchInput:
      type: object
      required: [clinicId, name, address]
      properties:
        slug: { type: string, minLength: 3, maxLength: 63 }
        email: { type: [string, "null"], format: email }
        inheritEmail: { type: boolean }
        inheritPhone: { type: boolean }
        openingHours: { type: [array, "null"], maxItems: 28, description: Null or absent is legacy unrestricted; empty array means closed all days., items: { $ref: "#/components/schemas/OpeningHour" } }
        clinicId: { type: string }
        name: { type: string, minLength: 1 }
        address: { type: string }
        city: { type: string }
        state: { type: string }
        pincode: { type: string }
        phone: { type: [string, "null"] }
        timezone: { type: string, default: Asia/Kolkata }
        status: { $ref: "#/components/schemas/RecordStatus" }
    Branch:
      allOf:
        - { $ref: "#/components/schemas/BranchInput" }
        - type: object
          required: [id, code, status, createdAt]
          properties:
            dateFormat: { $ref: "#/components/schemas/ClinicDateFormat" }
            timeFormat: { $ref: "#/components/schemas/ClinicTimeFormat" }
            linkedSchedule: { $ref: "#/components/schemas/LinkedSchedule" }
            effectiveEmail: { type: [string, "null"] }
            effectivePhone: { type: [string, "null"] }
            id: { type: string }
            code: { type: string }
            clinicName: { type: string }
            createdAt: { type: [string, "null"], format: date-time }
    DoctorInput:
      type: object
      required: [fullName, email]
      properties:
        ownerAdminId: { type: string, description: Backward-compatible only. If supplied it must match the managing admin derived by the server. }
        fullName: { type: string, minLength: 1 }
        email: { type: string, format: email }
        mobile: { type: string }
        photoUrl: { type: string }
        gender: { type: string }
        dateOfBirth: { type: string, format: date }
        specializationId: { type: string }
        qualificationIds: { type: array, items: { type: string } }
        registrationNumber: { type: string }
        experienceYears: { type: integer, minimum: 0 }
        about: { type: string }
        consultationFee: { type: number, minimum: 0 }
        languages: { type: array, items: { type: string } }
        clinicIds: { type: array, minItems: 1, items: { type: string }, description: Required by the server for doctor assignment. }
        branchIds: { type: array, items: { type: string } }
        status: { $ref: "#/components/schemas/RecordStatus" }
    Doctor:
      allOf:
        - { $ref: "#/components/schemas/DoctorInput" }
        - type: object
          required: [id, userId, code, ownerAdminId, managingAdminId, managingAdminName, invitationStatus, status, createdAt]
          properties:
            id: { type: string }
            userId: { type: string }
            code: { type: string }
            managingAdminId: { type: string }
            managingAdminName: { type: string }
            invitationStatus: { type: string, enum: [sent, failed, notRequired] }
            createdAt: { type: [string, "null"], format: date-time }
            specializationName: { type: string }
            qualificationNames: { type: array, items: { type: string } }
...